[formrecognizer] model_id param validation (#11569)
* check if model id is None or empty

* update for copy methods

+++ sdk/formrecognizer/azure-ai-formrecognizer/azure/ai/formrecognizer/_form_training_client.py
@@ -171,6 +171,9 @@ class FormTrainingClient(object):
                 :caption: Delete a custom model.
         """
 
+        if not model_id:
+            raise ValueError("model_id cannot be None or empty.")
+
         self._client.delete_custom_model(
             model_id=model_id,
             error_map=error_map,
@@ -244,6 +247,10 @@ class FormTrainingClient(object):
                 :dedent: 8
                 :caption: Get a custom model with a model ID.
         """
+
+        if not model_id:
+            raise ValueError("model_id cannot be None or empty.")
+
         response = self._client.get_custom_model(model_id=model_id, include_keys=True, error_map=error_map, **kwargs)
         return CustomFormModel._from_generated(response)
 
@@ -316,6 +323,9 @@ class FormTrainingClient(object):
                 :dedent: 8
                 :caption: Copy a model from the source resource to the target resource
         """
+
+        if not model_id:
+            raise ValueError("model_id cannot be None or empty.")
 
         polling_interval = kwargs.pop("polling_interval", POLLING_INTERVAL)
